修改实体bean的数据类型

+++ src/main/java/com/afd/model/product/AttrAttrValue.java
@@ -5,63 +5,64 @@ import java.io.Serializable;
 public class AttrAttrValue implements Serializable {
 	private static final long serialVersionUID = 4644425131428917414L;
 	
-    private Integer aAvId;
+    private Long aAvId;
 
-    private Integer attrId;
+    private Long attrId;
 
-    private Integer attrValueId;
+    private Long attrValueId;
 
-    private Short displayOrder;
+    private Integer displayOrder;
 
-    private Integer pAAvId;
+    private Long pAAvId;
 
     private Boolean hasSub;
 
-    public Integer getaAvId() {
-        return aAvId;
-    }
+	public Long getaAvId() {
+		return aAvId;
+	}
 
-    public void setaAvId(Integer aAvId) {
-        this.aAvId = aAvId;
-    }
+	public void setaAvId(Long aAvId) {
+		this.aAvId = aAvId;
+	}
 
-    public Integer getAttrId() {
-        return attrId;
-    }
+	public Long getAttrId() {
+		return attrId;
+	}
 
-    public void setAttrId(Integer attrId) {
-        this.attrId = attrId;
-    }
+	public void setAttrId(Long attrId) {
+		this.attrId = attrId;
+	}
 
-    public Integer getAttrValueId() {
-        return attrValueId;
-    }
+	public Long getAttrValueId() {
+		return attrValueId;
+	}
 
-    public void setAttrValueId(Integer attrValueId) {
-        this.attrValueId = attrValueId;
-    }
+	public void setAttrValueId(Long attrValueId) {
+		this.attrValueId = attrValueId;
+	}
 
-    public Short getDisplayOrder() {
-        return displayOrder;
-    }
+	public Integer getDisplayOrder() {
+		return displayOrder;
+	}
 
-    public void setDisplayOrder(Short displayOrder) {
-        this.displayOrder = displayOrder;
-    }
+	public void setDisplayOrder(Integer displayOrder) {
+		this.displayOrder = displayOrder;
+	}
 
-    public Integer getpAAvId() {
-        return pAAvId;
-    }
+	public Long getpAAvId() {
+		return pAAvId;
+	}
 
-    public void setpAAvId(Integer pAAvId) {
-        this.pAAvId = pAAvId;
-    }
+	public void setpAAvId(Long pAAvId) {
+		this.pAAvId = pAAvId;
+	}
 
-    public Boolean getHasSub() {
-        return hasSub;
-    }
+	public Boolean getHasSub() {
+		return hasSub;
+	}
 
-    public void setHasSub(Boolean hasSub) {
-        this.hasSub = hasSub;
-    }
+	public void setHasSub(Boolean hasSub) {
+		this.hasSub = hasSub;
+	}
+
 }
